Rewrite the C example programs against the typed draw ABI
