--- hegel-c/examples/invalid_schema.c
@@ -1,86 +1,0 @@
-/*
- * invalid_schema.c — regression demo for the hegel-java SIGABRT report.
- *
- * A plausible-but-wrong schema type (`{"type":"ipv4"}`) used to
- * `panic!("Unknown schema type")` inside the engine. Because hegel_generate
- * is `extern "C"`, that panic crossed the FFI boundary and aborted the whole
- * host process (SIGABRT) — it killed an embedding JVM.
- *
- * This program feeds such a schema to hegel_generate and asserts it returns
- * HEGEL_E_INVALID_ARG with a diagnostic, then exits 0. The interesting case
- * is the *statically linked* build under `panic = "abort"` (see
- * `just c-test-abort`): if any panic were reachable here, the process would
- * abort and this example would exit non-zero, failing the run.
- *
- * Build (from this directory, after `cargo build -p hegeltest-c --release`):
- *
- *   cc -o invalid_schema invalid_schema.c \
- *      -I../include \
- *      ../../target/release/libhegel.a -lpthread -ldl -lm -lrt
- *
- * Run:
- *   ./invalid_schema
- */
-
-#include <stdio.h>
-#include <stdint.h>
-#include <string.h>
-
-#include "hegel.h"
-#include "hegel_check.h"
-
-/*
- * Hand-rolled CBOR encoding of { "type": "ipv4" }:
- *   map(1) = 0xA1, text("type") = 0x64 + 4 bytes, text("ipv4") = 0x64 + 4 bytes.
- */
-static const uint8_t IPV4_TYPE_SCHEMA[] = {
-    0xA1,                            /* map(1) */
-    0x64, 't', 'y', 'p', 'e',        /* "type" */
-    0x64, 'i', 'p', 'v', '4',        /* "ipv4" */
-};
-
-int main(void) {
-    hegel_context_t *ctx = hegel_context_new();
-
-    hegel_settings_t *s;
-    HEGEL_CHECK(hegel_settings_new, ctx, &s);
-    HEGEL_CHECK(hegel_settings_set_test_cases, ctx, s, 1);
-    HEGEL_CHECK(hegel_settings_set_database, ctx, s, "");
-    HEGEL_CHECK(hegel_settings_set_derandomize, ctx, s, true);
-    HEGEL_CHECK(hegel_settings_set_seed, ctx, s, 1, true);
-
-    hegel_run_t *run;
-    HEGEL_CHECK(hegel_run_start, ctx, s, &run);
-
-    hegel_test_case_t *tc;
-    HEGEL_CHECK(hegel_next_test_case, ctx, run, &tc);
-    if (!tc) {
-        fprintf(stderr, "expected a test case\n");
-        return 1;
-    }
-
-    const uint8_t *value;
-    size_t value_len;
-    hegel_result_t rc = hegel_generate(ctx, tc, IPV4_TYPE_SCHEMA, sizeof(IPV4_TYPE_SCHEMA), &value, &value_len);
-
-    int ok = 1;
-    if (rc != HEGEL_E_INVALID_ARG) {
-        fprintf(stderr, "expected HEGEL_E_INVALID_ARG (%d), got rc=%d\n",
-                HEGEL_E_INVALID_ARG, rc);
-        ok = 0;
-    }
-    const char *err = hegel_context_last_error(ctx);
-    if (err[0] == '\0') {
-        fprintf(stderr, "expected a diagnostic message for the invalid schema\n");
-        ok = 0;
-    } else {
-        printf("invalid schema correctly rejected: rc=%d, message=\"%s\"\n", rc, err);
-    }
-
-    HEGEL_CHECK(hegel_mark_complete, ctx, tc, HEGEL_STATUS_INVALID, NULL);
-    HEGEL_CHECK(hegel_test_case_free, ctx, tc);
-    HEGEL_CHECK(hegel_run_free, ctx, run);
-    HEGEL_CHECK(hegel_settings_free, ctx, s);
-    HEGEL_CHECK(hegel_context_free, ctx);
-    return ok ? 0 : 1;
-}
